NXP-22961: Move logic to ExceptionUtils
